requirements.txt
llama-index-core>=0.10.0
llama-index-llms-vertex
llama-index-embeddings-vertex
llama-index-embeddings-huggingface
google-cloud-aiplatform
llama-index-vector-stores-pinecone
pinecone>=3.0.0
pymupdf
streamlit>=1.30.0
watchdog
python-dotenv
google-auth
docx2txt
python-pptx
llama-index-readers-file
transformers
sentence-transformers
fpdf